Refactor to remove unnecessary fragment
