+++ app/src/main/res/layout/activity_shuttle_stop.xml
